requirements.txt
streamlit
ollama
googlesearch-python
trafilatura